# speedup norm on AbstractArray{<:Number}