--- a/stdlib/LinearAlgebra/test/dense.jl
+++ b/stdlib/LinearAlgebra/test/dense.jl
@@ -2,6 +2,7 @@
 
 module TestDense
 
+using InteractiveUtils
 using Test, LinearAlgebra, Random
 using LinearAlgebra: BlasComplex, BlasFloat, BlasReal
 
@@ -330,6 +331,17 @@ end
 
                     # norm
                     for p in (-Inf, Inf, (-2:3)...)
+                        if norm(A, p) != norm(vec(A), p)
+                            @show A
+                            @show typeof(A)
+                            @show size(A)
+                            @show p
+                            @show InteractiveUtils.@which norm(A,p)
+                            @show InteractiveUtils.@which norm(vec(A),p)
+                            @show norm(A, p)
+                            @show norm(vec(A), p)
+                            error()
+                        end
                         @test norm(A, p) == norm(vec(A), p)
                     end
                 end