--- a/stdlib/LinearAlgebra/test/dense.jl
+++ b/stdlib/LinearAlgebra/test/dense.jl
@@ -338,8 +338,12 @@ end
                             @show p
                             @show InteractiveUtils.@which norm(A,p)
                             @show InteractiveUtils.@which norm(vec(A),p)
+                            @show InteractiveUtils.@which LinearAlgebra.norm2(A)
+                            @show InteractiveUtils.@which LinearAlgebra.norm2(vec(A))
                             @show norm(A, p)
                             @show norm(vec(A), p)
+                            @show LinearAlgebra.norm2(A)
+                            @show LinearAlgebra.norm2(vec(A))
                             error()
                         end
                         @test norm(A, p) == norm(vec(A), p)