This gives a nice speedup for generic arrays.
## PR
```julia
julia> arr = randn(10^5); oarr = OffsetArray(arr, eachindex(arr)); @btime norm(arr); @btime norm(oarr)
  29.836 μs (1 allocation: 16 bytes)
  73.266 μs (1 allocation: 16 bytes)
```
## Baseline
```julia
julia> arr = randn(10^5); oarr = OffsetArray(arr, eachindex(arr)); @btime norm(arr); @btime norm(oarr)

  29.833 μs (1 allocation: 16 bytes)
  254.549 μs (1 allocation: 16 bytes)
```