@DilumAluthge 2025-01-12
We have moved the LinearAlgebra stdlib to an external repo: https://github.com/JuliaLang/LinearAlgebra.jl

@jw3126 If you think that this PR is still relevant, please open a new PR on the [LinearAlgebra.jl](https://github.com/JuliaLang/LinearAlgebra.jl) repo.